Laying the foundation for automated tests of the new-style checkpoint-restart functionallity.
